\name{FG90}
\alias{FG90}
\docType{data}
\title{Pedigree, adapted from Table 1 in Fernando & Grossman (1990)
}

\description{An example pedigree}

\usage{data(FG90)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{id}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}}
    \item{\code{dam}}{a factor with levels \code{2} \code{4} \code{6}}
    \item{\code{sire}}{a factor with levels \code{1} \code{3} \code{5}}
    \item{\code{sex}}{a factor with levels \code{0} \code{1}}
  }
}


\source{
Fernando, R.L. & M. Grossman. 1990. Genetic evaluation with autosomal and X-chromosomal inheritance. Theoretical and Applied Genetics 80:75-80.
}

\keyword{datasets}
