\name{IBD}
\alias{IBD}
\title{
Identical by descent}
\description{
determines if two individuals are IBD, given their genotypes}
\usage{
IBD(pair, genotypes, n)
}
\arguments{
  \item{pair}{
which two individuals}
  \item{genotypes}{
each individuals' genotype}
  \item{n}{
number of individuals in the population to index the second haploid complement of genes}
}
\author{\email{matthewwolak@gmail.com}
}

