\name{IBD2}
\alias{IBD2}
\title{
Identity by descent helper function}
\description{
applies \code{IBD} over a set of pairs}
\usage{
IBD2(pairs, genos, n)
}
\arguments{
  \item{pairs }{
see \code{IBD}}
  \item{genos }{
see \code{IBD}}
  \item{n }{
see \code{IBD}}
}
\author{\email{matthewwolak@gmail.com}
}

