% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeM.R
\name{makeM}
\alias{makeM}
\title{Creates the (additive) mutational effects relationship matrix}
\usage{
makeM(pedigree)
}
\arguments{
\item{pedigree}{A pedigree where the columns are ordered ID, Dam, Sire}
}
\value{
Returns M, or the mutational effects relationship matrix, in sparse 
  matrix form.
}
\description{
This returns the (additive) mutational effects relationship matrix in sparse
  matrix format.
}
\details{
Missing parents (e.g., base population) should be denoted by either 'NA',
'0', or '*'.

See function \code{\link{makeMinv}} for directly obtaining the inverse of
the (additive) mutational effects genetic relationship matrix.
}
\examples{

 makeM(Mrode2)

}
\seealso{
\code{\link{makeA}}, \code{\link{makeS}}
}
\author{
\email{matthewwolak@gmail.com}
}
