% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMinv.R
\name{makeMinv}
\alias{makeMinv}
\title{Creates the inverse (additive) mutational effects relationship matrix}
\usage{
makeMinv(pedigree, ...)
}
\arguments{
\item{pedigree}{A pedigree where the columns are ordered ID, Dam, Sire}

\item{\dots}{Arguments to be passed to methods}
}
\value{
a \code{list}:
  \describe{
    \item{Minv }{the inverse of the (additive) mutational effects
      relationship matrix in sparse matrix form}
    \item{listAinv }{the three column list of the non-zero elements for the 
      inverse of the (additive) mutational effects relationship matrix.
      \code{attr(*, "rowNames")} links the integer for rows/columns to the ID
      column from the pedigree.}
    \item{f }{the individual coefficients of inbreeding for each individual 
      in the pedigree (matches the order of the first/ID column of the
      pedigree).}
    \item{logDet }{the log determinant of the M matrix}
    \item{dii }{the (non-zero) elements of the diagonal D matrix of the M=TDT'
      decomposition. Contains the variance of Mendelian sampling. Matches
      the order of the first/ID column of the pedigree.} 
  }
}
\description{
This returns the inverse of the numerator relationship matrix (inverse
additive genetic relatedness matrix). It can also be used to obtain
coefficients of inbreeding for the pedigreed population.
}
\details{
Missing parents (e.g., base population) should be denoted by either 'NA',
'0', or '*'.

The functions implement an adaptation of the Meuwissen and Luo (1992)
algorithm (particularly, following the description of the algorithm in
Mrode 2005) with some code borrowed from the \code{inverseA} function by
Jarrod Hadfield in the \code{MCMCglmm} package.
}
\examples{

 ##  Example pedigree from Wray 1990
   Wray90 <- data.frame(id = seq(8),
dam = c(NA, NA, 1, 1, 4, 4, 6, 7),
sire = c(NA, NA, 2, 2, 3, 2, 5, 5),
time = c(0, 0, 1, 1, 2, 2, 3, 4))
   makeMinv(Mrode2)
 

}
\references{
Fikse, F. 2009. Fuzzy classification of phantom parent groups in
an animal model. Genetics Selection Evolution 41:42.

Meuwissen, T.H.E & Luo, Z. 1992. Computing inbreeding coefficients in large
populations. Genetics, Selection, Evolution. 24:305-313.

Mrode, R.A. 2005. Linear Models for the Prediction of Animal Breeding
Values, 2nd ed.  Cambridge, MA: CABI Publishing.

Quaas, R.L. 1988. Additive genetic model with groups and relationships.
Journal of Dairy Science. 71:1338-1345.

VanRaden, P.M. 1992. Accounting for inbreeding and crossbreeding in genetic
evaluation of large populations. Journal of Dairy Science. 75:3136-3144.
}
\author{
\email{matthewwolak@gmail.com}
}
