\name{Mrode2}
\alias{Mrode2}
\docType{data}
\title{Pedigree from Table 2.1 of Mrode (2005)
}
\description{
An example pedigree
}
\usage{data(Mrode2)}
\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{id} }{a numeric vector}
    \item{\code{dam} }{a numeric vector}
    \item{\code{sire} }{a numeric vector}
  }
}
\source{Mrode, R.A. 2005. Linear Models for the Prediction of Animal Breeding Values, 2nd ed.  Cambridge, MA: CABI Publishing.
}
\keyword{datasets}
