% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadiv-datasets.R
\docType{data}
\name{Mrode3}
\alias{Mrode3}
\title{Pedigree, from chapter 3 of Mrode (2005) with genetic groups and a trait column}
\format{
A \code{data.frame} with 10 observations on the following 8 variables:
  \describe{
    \item{calf }{a factor with levels indicating the unique genetic groups 
      and individuals}
    \item{dam }{a numeric vector of maternal identities}
    \item{sire }{a numeric vector of paternal identities}
    \item{damGG }{a factor of maternal identities with genetic groups 
      inserted instead of \code{NA}}
    \item{sireGG }{a factor of paternal identities with genetic groups 
      inserted instead of \code{NA}}
    \item{sex }{a factor with levels \code{female} \code{male}}
    \item{WWG }{a numeric vector of pre-weaning weight gain (kg) for five 
      beef calves}
  }
}
\source{
Mrode, R.A. 2005. Linear Models for the Prediction of Animal
Breeding Values, 2nd ed.  Cambridge, MA: CABI Publishing.
}
\usage{
Mrode3
}
\description{
Pedigree, from chapter 3 of Mrode (2005) with genetic groups and a trait column
}
\examples{
  data(Mrode3)
  str(Mrode3)
}
\keyword{datasets}
