\name{Mrode9}
\alias{Mrode9}
\docType{data}
\title{Pedigree, adapted from example 9.1 of Mrode (2005)}
\description{
An example pedigree}
\usage{data(Mrode9)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{pig} }{a numeric vector}
    \item{\code{dam} }{a numeric vector}
    \item{\code{sire} }{a numeric vector}
  }
}
\source{Mrode, R.A. 2005. Linear Models for the Prediction of Animal Breeding Values, 2nd ed.  Cambridge, MA: CABI Publishing.
}
\keyword{datasets}
