% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remlSupport.R
\name{pin-deprecated}
\alias{pin-deprecated}
\title{Approximate standard errors for linear functions of variance components}
\usage{
pin(object, transform)
}
\arguments{
\item{object}{A list with at least the following elements: \code{gammas},
\code{gammas.type}, and \code{ai} from a REML mixed model}

\item{transform}{A formula specifying the linear transformation of variance
components to conduct}
}
\value{
A \code{data.frame} with row names corresponding to the operator on 
  the left hand side of the \code{transform} formula and the entries 
  corresponding to the \code{Estimate} and approximate \code{SE} of the 
  linear transformation.
}
\description{
This function is Deprecated and will be removed in a future version. The pin
function works with an asreml-R version 3 model object. Since ASReml has
updated to version 4, they have changed their model output. ASReml has also
provided their own \code{vpredict} function that does (for asreml v4 model
objects) what \code{pin} did for asreml v3 model objects.
}
\details{
This function is similar to the pin calculations performed by the standalone
ASReml.  This function, written by Ian White, applies the delta method for
the estimation of approximate standard errors on linear functions of
variance components from a REML mixed model

Object is intended to be an asreml-R model output.

The formula can use \code{V1,..., Vn} to specify any one of the \code{n}
variance components.  These should be in the same order as they are in the
object (e.g., see the row order of \code{summary(object)$varcomp} for
asreml-R models.
}
\seealso{
See Also \code{\link{nadiv-deprecated}},  \code{\link{aiCI}},
  \code{\link{aiFun}}
}
\author{
Ian White
}
\keyword{internal}
