% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadiv-datasets.R
\docType{data}
\name{Q1988}
\alias{Q1988}
\title{Pedigree, adapted from Quaas (1988) equation [5]}
\format{
An object of class \code{data.frame} with 11 rows and 8 columns.
}
\source{
Quaas, R.L. 1988. Additive genetic model with groups and
relationships. Journal of Dairy Science 71:1338-1345.
}
\usage{
Q1988
}
\description{
An example pedigree with genetic groups
}
\details{
A \code{data.frame} with 11 observations on the following 8 variables:
  \describe{
    \item{id }{a factor with levels indicating the unique individuals 
      (including phantom parents) and genetic groups}
    \item{dam }{a factor of observed maternal identities}
    \item{sire }{a factor vector of observed paternal identities}
    \item{damGG }{a factor of maternal identities with genetic groups 
      inserted instead of \code{NA}}
    \item{sireGG }{a factor of paternal identities with genetic groups 
      inserted instead of \code{NA}}
    \item{phantomDam }{a factor of maternal identities with phantom parents 
      inserted instead of \code{NA}}
    \item{phantomSire }{a factor of paternal identities with phantom parents 
      inserted instead of \code{NA}}
    \item{group }{a factor of genetic groups to which each phantom parent
      belongs}
  }
}
\examples{
  data(Q1988)
  str(Q1988)
}
\keyword{datasets}
