\name{SEfun}
\alias{SEfun}
\title{
Estimate of the standard error for a simulated coefficient of fraternity}
\description{
The standard error for coefficients of fraternity that are derived from simulation (\code{\link{makeDsim}}) as described in the appendix to Ovaskainen et al. (2008).}
\usage{
SEfun(coef.frat, N)
}
\arguments{
  \item{coef.frat}{
the simulated coefficient of fraternity}
  \item{N}{
the number of times the simulation was conducted}
}
\references{
Ovaskainen, O., Cano, J.M., & Merila, J. 2008. A Bayesian framework for comparative quantitative genetics. Proceedings of the Royal Society B 275, 669-678.
}
\author{\email{matthewwolak@gmail.com}
}

