% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPed.R
\name{simPedDFC}
\alias{simPedDFC}
\title{Double first cousin pedigree construction}
\usage{
simPedDFC(U, gpn = 4, fsn = 4, s = 2, fws = 2, prefix = NULL)
}
\arguments{
\item{U}{An integer number of units or blocks for the design}

\item{gpn}{Number of grandparent pairs per unit in the generation 0 (GP)
(must be >= 2). Equals the number of full-sib families in generation 1 (P).}

\item{fsn}{Number of offspring in each full-sib family of generations 1 and 2
(P and F1 - must be an even number >= 4).}

\item{s}{Number of sires per full-sib family in generation 1 (P - must be >=2)}

\item{fws}{Number of generation 1 (P) families with sires in each unit.
Together, with \code{s}, sets up how cousins and double first cousins are
produced}

\item{prefix}{Optional prefix to add to every identity}
}
\value{
A \code{data.frame} with columns corresponding to: id, dam, sire,
  and sex. Sex is \code{M} for males and \code{F} for females.
}
\description{
Simulates a pedigree for the \dQuote{double first cousin} mating design
(Fairbairn and Roff 2006).
}
\details{
This is an adaption to a half-sib breeding design which also produces first
cousins and double first cousins. Double first cousins are produced by
mating two brothers to two sisters (the offspring of the resulting two
families are double first cousins with one another). This is described in
Fairbairn and Roff (2006) as being particularly effective for separating
autosomal additive genetic variance from sex chromosomal additive genetic
variance. It is also amenable to estimating dominance variance, however, it
still has difficulty separating dominance variance from common maternal
environmental variance (Meyer 2008).

For a given unit of the design (\code{U} total), \code{2*gpn} 0-generation 
(grandparental or GP) individuals are created and paired to make \code{gpn}
full-sib families. Then the first \code{fws} families are each allocated
\code{s} males/sires and \code{s*(fws-1)} females/dams in the 1 (parental or P)
generation. The remaining (\code{gpn-fws}) families (only when: 
\code{gpn > fws}) are assigned \code{s*fws} females/dams. If
\code{fsn > (s*fws)}, the remaining generation 1 (P) individuals in each 
full-sib family (\code{fsn - (s*fws)}) are allocated to each family with
equal numbers of females and males [this allows for more individuals to be
phenotyped in generation 1 (P) than are used to produce generation 2 (F1)].
Generation 2 (F1) is then assigned, based on the mating design in Fairbairn
and Roff (2006) - essentially each sire [of the \code{s} per full-sib family
in generation 1 (P)] is mated to a female from each of the other \code{gpn-1}
full-sib families to produce \code{fsn} offspring (with equal numbers of
females and males).
}
\examples{

  DFC1 <- simPedDFC(U = 1, gpn = 2, fsn = 4, s = 2, fws = 2)

}
\references{
Fairbairn, D.J. and D.A. Roff. 2006.  The quantitative genetics
of sexual dimorphism: assessing the importance of sex-linkage. Heredity
97:319-328.

Meyer, K. 2008.  Likelihood calculations to evaluate experimental designs to
estimate genetic variances. Heredity 101:212-221.
}
\seealso{
\code{\link{simPedHS}}, \code{\link{warcolak}}
}
\author{
\email{matthewwolak@gmail.com}
}
