\name{wrap_DFC}
\alias{wrap_DFC}
\title{
Double First Cousin wrapper function}
\description{
Used in the \code{findDFC} function}
\usage{
wrap_DFC(x, grandparents)
}
\arguments{
  \item{x }{
sequence}
  \item{grandparents }{
list of all individuals and their grandparents}
}
\details{
Used to wrap the \code{DFC} function inside of \code{findDFC}}
\author{\email{matthewwolak@gmail.com}
}

