\name{wrap_dij}
\alias{wrap_dij}
\title{Wrapper function for \code{makeD}
}

\description{
used inside \code{makeD}}

\usage{
wrap_dij(x, parents, relatedness, N)
}
\arguments{
  \item{x }{sequence}
  \item{parents }{parents of both individuals, i and j}
  \item{relatedness }{list form of relatedness matrix}
  \item{N }{number of individuals in the pedigree}
}

\author{\email{matthewwolak@gmail.com}
}

