% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiCI.R
\name{aiCI}
\alias{aiCI}
\title{Confidence Intervals for Variance Components}
\usage{
aiCI(asr.model, Dimnames = NULL, alpha = 0.05)
}
\arguments{
\item{asr.model}{Object from a call to \code{asreml}}

\item{Dimnames}{A vector of characters if names are desired for the output.
If not specified, the default labels from the \code{asreml} object will be
used.}

\item{alpha}{A numeric value indicating the level of Type I error for
constructing the Confidence Intervals.}
}
\value{
A \code{matrix} is returned with a row for each variance component. 
  The three columns correspond to the Lower Confidence Limit, estimate from
  the \code{asreml} model, and Upper Confidence Limit for each variance 
  component.
}
\description{
Produces the 1-alpha Upper and Lower Confidence Limits for the variance
components in an ASReml-R model.
}
\details{
Variances from the inverse of the Average Information matrix of an ASReml
model are translated according to the \code{\link{varTrans}} function and
used in constructing the 1-alpha Confidence Interval.
}
\note{
The vector of \code{Dimnames} should match the same order of variance
components specified in the model.
}
\examples{

  \dontrun{
    library(asreml)
    ginvA <- ainverse(warcolak)
    ginvD <- makeD(warcolak[, 1:3])$listDinv
      attr(ginvD, "rowNames") <- as.character(warcolak[, 1])
      attr(ginvD, "INVERSE") <- TRUE
    warcolak$IDD <- warcolak$ID
    warcolak.mod <- asreml(trait1 ~ sex,
     random = ~ vm(ID, ginvA) + vm(IDD, ginvD), 
	data = warcolak) 
    summary(warcolak.mod)$varcomp
    aiCI(warcolak.mod)    
   }

}
\seealso{
\code{\link{aiFun}} \code{\link{proLik}}
}
\author{
\email{matthewwolak@gmail.com}
}
