% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadiv-datasets.R
\docType{data}
\name{FG90}
\alias{FG90}
\title{Pedigree, adapted from Table 1 in Fernando & Grossman (1990)}
\format{
A \code{data.frame} with 8 observations on the following 4 variables:
  \describe{
    \item{id }{a factor with levels \code{1} \code{2} \code{3} \code{4} 
      \code{5} \code{6} \code{7} \code{8}}
    \item{dam }{a factor with levels \code{2} \code{4} \code{6}}
    \item{sire }{a factor with levels \code{1} \code{3} \code{5}}
    \item{sex }{a factor with levels \code{0} \code{1}}
  }
}
\source{
Fernando, R.L. & M. Grossman. 1990. Genetic evaluation with
autosomal and X-chromosomal inheritance. Theoretical and Applied Genetics
80:75-80.
}
\usage{
FG90
}
\description{
Pedigree, adapted from Table 1 in Fernando & Grossman (1990)
}
\examples{
  data(FG90)
  str(FG90)
}
\keyword{datasets}
