% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeA.R
\name{makeA}
\alias{makeA}
\title{Creates the additive genetic relationship matrix}
\usage{
makeA(pedigree)
}
\arguments{
\item{pedigree}{A pedigree where the columns are ordered ID, Dam, Sire}
}
\value{
Returns A, or the numerator relationship matrix, in sparse 
  matrix form.
}
\description{
This returns the additive relationship matrix in sparse matrix format.
}
\details{
Missing parents (e.g., base population) should be denoted by either 'NA',
'0', or '*'.

Used as a support function to \code{\link{makeD}}.

See function \code{\link{makeAinv}} for directly obtaining the inverse of
the additive genetic relationship matrix.
}
\examples{

 makeA(Mrode2)

}
\seealso{
\code{\link{makeD}}, \code{\link{makeS}}
}
\author{
\email{matthewwolak@gmail.com}
}
