% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadiv-datasets.R
\docType{data}
\name{Mrode2}
\alias{Mrode2}
\title{Pedigree from Table 2.1 of Mrode (2005)}
\format{
A \code{data.frame} with 6 observations on the following 3 variables:
  \describe{
    \item{id }{a numeric vector}
    \item{dam }{a numeric vector}
    \item{sire }{a numeric vector}
  }
}
\source{
Mrode, R.A. 2005. Linear Models for the Prediction of Animal
Breeding Values, 2nd ed.  Cambridge, MA: CABI Publishing.
}
\usage{
Mrode2
}
\description{
Pedigree from Table 2.1 of Mrode (2005)
}
\examples{
  str(Mrode2)
}
\keyword{datasets}
