% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadiv-datasets.R
\docType{data}
\name{Mrode9}
\alias{Mrode9}
\title{Pedigree, adapted from example 9.1 of Mrode (2005)}
\format{
A \code{data.frame} with 12 observations on the following 3 variables:
  \describe{
    \item{pig }{a numeric vector}
    \item{dam }{a numeric vector}
    \item{sire }{a numeric vector}
  }
}
\source{
Mrode, R.A. 2005. Linear Models for the Prediction of Animal
Breeding Values, 2nd ed.  Cambridge, MA: CABI Publishing.
}
\usage{
Mrode9
}
\description{
Pedigree, adapted from example 9.1 of Mrode (2005)
}
\examples{
  data(Mrode9)
  str(Mrode9)
}
\keyword{datasets}
