\name{nadiv-package}
\alias{nadiv-package}
\alias{nadiv}
\docType{package}
\title{Functions to construct (non)additive genetic relatedness matrices and facilitate animal model analyses
}
\description{Constructs (non)additive genetic relationship matrices, and their inverses, from a pedigree to be used in linear mixed effect models (A.K.A. the 'animal model').  Also includes other functions to facilitate the use of animal models.
}
\details{
\tabular{ll}{
Package: \tab nadiv\cr
Type: \tab Package\cr
Version: \tab 2.13.2\cr
Date: \tab 2014-06-20\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Matthew Wolak \email{matthewwolak@gmail.com}
}

