% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadiv.R
\docType{package}
\name{nadiv-package}
\alias{nadiv-package}
\alias{nadiv}
\title{(Non)Additive Genetic Relatedness Matrices in Animal Model Analyses}
\description{
Constructs (non)additive genetic relationship matrices, and their inverses,
from a pedigree to be used in linear mixed effect models (A.K.A. the 'animal
model'). Also includes other functions to facilitate the use of animal
models. Some functions have been created to be used in conjunction with the
R package for ASReml software, which can be obtained upon purchase from
VSN international (<https://www.vsni.co.uk/software/asreml>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewwolak/nadiv}
  \item Report bugs at \url{https://github.com/matthewwolak/nadiv/issues}
}

}
\author{
Matthew Wolak \email{matthewwolak@gmail.com}
}
