\name{nadiv-package}
\alias{nadiv-package}
\alias{nadiv}
\docType{package}
\title{(Non)Additive Genetic Relatedness Matrices in Animal Model Analyses
}
\description{Constructs (non)additive genetic relationship matrices, and their inverses, from a pedigree to be used in linear mixed effect models (A.K.A. the 'animal model'). Also includes other functions to facilitate the use of animal models.
}
\details{
\tabular{ll}{
Package: \tab nadiv\cr
Type: \tab Package\cr
Version: \tab 2.13.3\cr
Date: \tab 2015-05-30\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Matthew Wolak \email{matthewwolak@gmail.com}
}

