\name{nadiv-package}
\alias{nadiv-package}
\alias{nadiv}
\docType{package}
\title{(Non)Additive Genetic Relatedness Matrices in Animal Model Analyses
}
\description{Constructs (non)additive genetic relationship matrices, and their inverses, from a pedigree to be used in linear mixed effect models (A.K.A. the 'animal model'). Also includes other functions to facilitate the use of animal models. Some functions have been created to be used in conjunction with the R package for 'ASReml' software, which can be obtained upon purchase from 'VSN' international (http://www.vsni.co.uk/software/asreml).
}
\details{
\tabular{ll}{
Package: \tab nadiv\cr
Type: \tab Package\cr
Version: \tab 2.15.0.4\cr
Date: \tab 2017-10-25\cr
URL: \tab http://github.com/matthewwolak/nadiv\cr
BugReports: \tab http://github.com/matthewwolak/nadiv/issues\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Matthew Wolak \email{matthewwolak@gmail.com}
}

